Allow adding two dependencies for an input if....
one dependency is a dependency on a previous timestep and the other
is a dependency on a scenario. This allows self-dependencies where
initial input data can come from a scenario.

Move ScenarioModel into smif.model.model to break circular import
between Model<->ScenarioModel.
